script/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.73.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.10.0
asttokens @ file:///croot/asttokens_1743630435401/work
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
black==25.12.0
bleach==6.3.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.3.0
colabcode==0.1.1
coloredlogs==15.0.1
comm @ file:///croot/comm_1709322850197/work
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy @ file:///home/task_176242086284533/conda-bld/debugpy_1762421636131/work
decorator @ file:///home/task_175734035162575/conda-bld/decorator_1757341232127/work
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.17.0
docx==0.2.4
durationpy==0.10
emoji==2.15.0
et_xmlfile==2.0.0
executing @ file:///home/task_175706124557094/conda-bld/executing_1757061259817/work
faiss-cpu==1.12.0
fastapi==0.121.1
fastjsonschema==2.21.2
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.61.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.33.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.4.0
importlib_resources==6.5.2
ipykernel @ file:///home/task_175758336750008/conda-bld/ipykernel_1757583382803/work
ipython @ file:///home/task_175932809916977/conda-bld/ipython_1759328496611/work
ipython_pygments_lexers @ file:///croot/ipython_pygments_lexers_1744753235686/work
ipywidgets==8.1.8
isoduration==20.11.0
jedi @ file:///croot/jedi_1733987392413/work
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client @ file:///croot/jupyter_client_1737570961872/work
jupyter_core @ file:///croot/jupyter_core_1751991368470/work
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kaggle==1.8.2
kagglesdk==0.1.13
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==1.0.8
langchain-community==0.0.38
langchain-core==1.0.7
langchain-groq==1.0.1
langchain-huggingface==1.0.1
langchain-text-splitters==0.0.2
langdetect==1.0.9
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langsmith==0.4.45
lark==1.3.1
librt==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline @ file:///home/task_175851944441289/conda-bld/matplotlib-inline_1758520101789/work
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.10.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.5
nltk==3.9.2
notebook==7.5.0
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.7.2
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.11.4
ormsgpack==1.12.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.0.2
pandocfilters==1.5.1
parso @ file:///croot/parso_1733963305961/work
pathspec==0.12.1
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect @ file:///home/task_176253519276066/conda-bld/pexpect_1762535937040/work
pillow==10.4.0
platformdirs @ file:///home/task_176235638439362/conda-bld/platformdirs_1762356487841/work
posthog==6.9.1
prometheus_client==0.23.1
prompt_toolkit @ file:///home/task_176174483353592/conda-bld/prompt-toolkit_1761744845973/work
propcache==0.4.1
protobuf==4.25.8
psutil @ file:///home/task_176189642143064/conda-bld/psutil_1761896460324/work
ptyprocess @ file:///opt/miniconda3/conda-bld/ptyprocess_1762424170819/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=3470be7f810474c8a2ecfcd6e02acc6aea8483ab595417fa4e336362a349933e
pure_eval @ file:///home/task_175706703828594/conda-bld/pure_eval_1757067053474/work
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments @ file:///home/task_176243133773609/conda-bld/pygments_1762431407413/work
pyngrok==7.5.0
pyparsing==3.2.5
pypdf==6.1.3
PyPDF2==3.0.1
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
PySocks==1.7.1
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-docx==1.2.0
python-dotenv==1.0.0
python-iso639==2025.2.18
python-json-logger==4.0.0
python-magic==0.4.27
python-slugify==8.0.4
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq @ file:///home/task_176237559798588/conda-bld/pyzmq_1762375611039/work
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
selenium==4.21.0
Send2Trash==1.8.3
sentence-transformers==5.1.2
shellingham==1.5.4
six @ file:///croot/six_1744271502820/work
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
stack_data @ file:///home/task_175706683579810/conda-bld/stack_data_1757067036897/work
starlette==0.49.3
streamlit==1.35.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado @ file:///croot/tornado_1748956929273/work
tqdm==4.67.1
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.57.1
trio==0.32.0
trio-websocket==0.12.2
triton==3.5.0
typer==0.20.0
typer-slim==0.20.0
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
tzdata==2025.2
unstructured==0.14.6
unstructured-client==0.42.3
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth @ file:///croot/wcwidth_1750352883074/work
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
wrapt==1.17.3
wsproto==1.2.0
xlrd==2.0.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
